模型可視化與驗證教材/requirements.txt
# 請確保安裝好 PyTorch (https://pytorch.org/get-started/locally/)
matplotlib
numpy
Pillow
opencv-python
ultralytics
grad-cam